refactor: change all Long ids to Integer

+++ src/main/java/se/ifmo/address/AddressController.java
@@ -8,7 +8,7 @@ import se.ifmo.common.AbstractCrudController;
 @RequestMapping("api/v1/address")
 public class AddressController extends AbstractCrudController<
         AddressDto,
-        Long,
+        Integer,
         AddressService
         >{
