+++ src/main/java/se/ifmo/person/PersonRepository.java
@@ -4,5 +4,5 @@ import org.springframework.stereotype.Repository;
 import se.ifmo.common.placemark.AbstractRepository;
 
 @Repository
-public interface PersonRepository extends AbstractRepository<Person, Long> {
+public interface PersonRepository extends AbstractRepository<Person, Integer> {
 }
